restructured for tests to be separate
